Bit of refactoring to merge code into single file
